Revert "Adapt to new TDLib"
This reverts commit 1e04543f5e04ecf02f78c61dcf68893a41be1169.

+++ app/src/main/java/org/thunderdog/challegram/data/MediaWrapper.java
@@ -1245,8 +1245,8 @@ public class MediaWrapper implements FileProgressComponent.SimpleListener, FileP
           twAlternativeHeader = textShort;
           text = Lang.getString(R.string.ProcessingMedia, textShort);
         } else {
-          long progressSize = fileProgress.getProgressSize();
-          long totalSize = fileProgress.getTotalSize();
+          int progressSize = fileProgress.getProgressSize();
+          int totalSize = fileProgress.getTotalSize();
           if (progressSize <= totalSize) {
             float ratio = (float) progressSize / (float) totalSize;
             text = Lang.getString(R.string.format_uploadStatus, (int) Math.floor(100f * ratio), textShort);
